requirements.txt
ccxt==4.3.57
matplotlib==3.9.1
newsapi-python==0.2.7
nltk==3.9.1
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
streamlit==1.37.0
tensorflow==2.17.0